+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Log the root directory in the output console.
 - Fix that the test run finished prematurely when the first scenario finished.
 - Log `world.attach` messages as test output.
+- Fix that the number of seconds the test step took were omitted - only the nanoseconds were counted.
 
 ## [0.7.3]